Test playerattack returns correct damage + Changed NPC HP to double f…

+++ src/main/java/npc/Albatross.java
@@ -39,7 +39,7 @@ public class Albatross extends EnemyNPC implements Killable {
     }
 
     @Override
-    public int getHP() {
+    public double getHP() {
         return 0;
     }
